add subspecs so I can import only Logs and Traces code
