+++ OpenTelemetry-Swift-Protocol-Exporter-Http.podspec
@@ -9,7 +9,6 @@ Pod::Spec.new do |spec|
   spec.authors = "OpenTelemetry Authors"
 
   spec.source = { :git => "https://github.com/open-telemetry/opentelemetry-swift.git", :tag => spec.version.to_s }
-  spec.source_files = "Sources/Exporters/OpenTelemetryProtocolHttp/**/*.swift"
 
   spec.swift_version = "5.10"
   spec.ios.deployment_target = "13.0"
@@ -17,11 +16,41 @@ Pod::Spec.new do |spec|
   spec.watchos.deployment_target = "6.0"
   spec.module_name = "OpenTelemetryProtocolExporterHttp"
 
-  spec.dependency 'OpenTelemetry-Swift-Api', spec.version.to_s
-  spec.dependency 'OpenTelemetry-Swift-Sdk', spec.version.to_s
-  spec.dependency 'OpenTelemetry-Swift-Protocol-Exporter-Common', spec.version.to_s
   spec.dependency 'OpenTelemetry-Swift-DataCompression', spec.version.to_s
   spec.dependency 'SwiftProtobuf', '~> 1.28'
   spec.pod_target_xcconfig = { "OTHER_SWIFT_FLAGS" => "-module-name OpenTelemetryProtocolExporterHttp -package-name opentelemetry_swift_exporter_http" }
+
+  spec.subspec 'Core' do |s|
+    s.dependency 'OpenTelemetry-Swift-Api/Core'
+    s.dependency 'OpenTelemetry-Swift-Sdk/Core'
+    spec.dependency 'OpenTelemetry-Swift-Protocol-Exporter-Common/Core'
+    s.source_files = 'Sources/Exporters/OpenTelemetryProtocolHttp/*.swift'
+  end
+
+  spec.subspec 'Metrics' do |s|
+    s.dependency 'OpenTelemetry-Swift-Api/Metrics'
+    s.dependency 'OpenTelemetry-Swift-Sdk/Metrics'
+    s.dependency 'OpenTelemetry-Swift-Protocol-Exporter-Http/Core'
+    s.source_files = 'Sources/Exporters/OpenTelemetryProtocolHttp/metric/*.swift'
+
+    s.subspec 'ExporterMetrics' do |sub|
+      sub.dependency 'OpenTelemetry-Swift-Protocol-Exporter-Common/ExporterMetrics'
+      sub.source_files = ['Sources/Exporters/OpenTelemetryProtocolHttp/metric/metrics/*.swift', 'Sources/Exporters/OpenTelemetryProtocolHttp/logs/metrics/*.swift', 'Sources/Exporters/OpenTelemetryProtocolHttp/trace/metrics/*.swift']
+    end
+  end
+
+  spec.subspec 'Logs' do |s|
+    s.dependency 'OpenTelemetry-Swift-Api/Logs'
+    s.dependency 'OpenTelemetry-Swift-Sdk/Logs'
+    s.dependency 'OpenTelemetry-Swift-Protocol-Exporter-Http/Core'
+    s.source_files = 'Sources/Exporters/OpenTelemetryProtocolHttp/logs/*.swift'
+  end
+
+  spec.subspec 'Trace' do |s|
+    s.dependency 'OpenTelemetry-Swift-Api/Trace'
+    s.dependency 'OpenTelemetry-Swift-Sdk/Trace'
+    s.dependency 'OpenTelemetry-Swift-Protocol-Exporter-Http/Core'
+    s.source_files = 'Sources/Exporters/OpenTelemetryProtocolHttp/trace/*.swift'
+  end
 
 end
