+++ OpenTelemetry-Swift-Sdk.podspec
@@ -9,12 +9,37 @@ Pod::Spec.new do |spec|
   spec.authors = "OpenTelemetry Authors"
 
   spec.source = { :git => "https://github.com/open-telemetry/opentelemetry-swift.git", :tag => spec.version.to_s }
-  spec.source_files = "Sources/OpenTelemetrySdk/**/*.swift"
 
   spec.swift_version = "5.10"
   spec.ios.deployment_target = "13.0"
   spec.tvos.deployment_target = "13.0"
   spec.watchos.deployment_target = "6.0"
   spec.module_name = "OpenTelemetrySdk"
-  spec.dependency 'OpenTelemetry-Swift-Api', spec.version.to_s
+
+  spec.subspec 'Core' do |s|
+    s.dependency 'OpenTelemetry-Swift-Api/Core'
+    s.source_files = ['Sources/OpenTelemetrySdk/*.swift', 'Sources/OpenTelemetrySdk/Common/**/*.swift', 'Sources/OpenTelemetrySdk/Internal/**/*.swift', 'Sources/OpenTelemetrySdk/Resources/**/*.swift']
+  end
+
+  spec.subspec 'Logs' do |s|
+    s.dependency 'OpenTelemetry-Swift-Sdk/Core'
+    s.source_files = 'Sources/OpenTelemetrySdk/Logs/**/*.swift'
+  end
+
+  spec.subspec 'Metrics' do |s|
+    s.dependency 'OpenTelemetry-Swift-Sdk/Core'
+    s.source_files = 'Sources/OpenTelemetrySdk/Metrics/**/*.swift'
+  end
+
+  spec.subspec 'Trace' do |s|
+    s.dependency 'OpenTelemetry-Swift-Sdk/Core'
+    s.source_files = 'Sources/OpenTelemetrySdk/Trace/**/*.swift'
+  end
+
+  spec.subspec 'TraceMetrics' do |s|
+    s.dependency 'OpenTelemetry-Swift-Sdk/Trace'
+    s.dependency 'OpenTelemetry-Swift-Sdk/Metrics'
+    s.source_files = 'Sources/OpenTelemetrySdk/TraceMetrics/**/*.swift'
+  end
+
 end
