+++ Sources/Exporters/OpenTelemetryProtocolHttp/metric/OltpHTTPMetricExporter.swift
@@ -20,7 +20,7 @@ public func defaultOltpHTTPMetricsEndpoint() -> URL {
 public class OtlpHttpMetricExporter: OtlpHttpExporterBase, MetricExporter {
   var pendingMetrics: [Metric] = []
   private let exporterLock = Lock()
-  private var exporterMetrics: ExporterMetrics?
+  internal var exporterMetrics: IExporterMetrics?
 
   override
   public init(endpoint: URL = defaultOltpHTTPMetricsEndpoint(),
@@ -31,28 +31,6 @@ public class OtlpHttpMetricExporter: OtlpHttpExporterBase, MetricExporter {
                config: config,
                useSession: useSession,
                envVarHeaders: envVarHeaders)
-  }
-
-  /// A `convenience` constructor to provide support for exporter metric using`StableMeterProvider` type
-  /// - Parameters:
-  ///    - endpoint: Exporter endpoint injected as dependency
-  ///    - config: Exporter configuration including type of exporter
-  ///    - meterProvider: Injected `StableMeterProvider` for metric
-  ///    - useSession: Overridden `URLSession` if any
-  ///    - envVarHeaders: Extra header key-values
-  public convenience init(endpoint: URL = defaultOltpHTTPMetricsEndpoint(),
-                          config: OtlpConfiguration = OtlpConfiguration(),
-                          meterProvider: StableMeterProvider,
-                          useSession: URLSession? = nil,
-                          envVarHeaders: [(String, String)]? = EnvVarHeaders.attributes) {
-    self.init(endpoint: endpoint, config: config, useSession: useSession,
-              envVarHeaders: envVarHeaders)
-    exporterMetrics = ExporterMetrics(type: "metric",
-                                      meterProvider: meterProvider,
-                                      exporterName: "otlp",
-                                      transportName: config.exportAsJson
-                                        ? ExporterMetrics.TransporterType.httpJson
-                                        : ExporterMetrics.TransporterType.grpc)
   }
 
   public func export(metrics: [Metric], shouldCancel: (() -> Bool)?)
