+++ Sources/Exporters/OpenTelemetryProtocolHttp/metric/StableOtlpHTTPMetricExporter.swift
@@ -22,7 +22,7 @@ public class StableOtlpHTTPMetricExporter: StableOtlpHTTPExporterBase, StableMet
 
   var pendingMetrics: [StableMetricData] = []
   private let exporterLock = Lock()
-  private var exporterMetrics: ExporterMetrics?
+  internal var exporterMetrics: IExporterMetrics?
 
   // MARK: - Init
 
@@ -37,38 +37,6 @@ public class StableOtlpHTTPMetricExporter: StableOtlpHTTPExporterBase, StableMet
 
     super.init(endpoint: endpoint, config: config, useSession: useSession,
                envVarHeaders: envVarHeaders)
-  }
-
-  /// A `convenience` constructor to provide support for exporter metric using`StableMeterProvider` type
-  /// - Parameters:
-  ///    - endpoint: Exporter endpoint injected as dependency
-  ///    - config: Exporter configuration including type of exporter
-  ///    - meterProvider: Injected `StableMeterProvider` for metric
-  ///    - aggregationTemporalitySelector: aggregator
-  ///    - defaultAggregationSelector: default aggregator
-  ///    - useSession: Overridden `URLSession` if any
-  ///    - envVarHeaders: Extra header key-values
-  public convenience init(endpoint: URL,
-                          config: OtlpConfiguration = OtlpConfiguration(),
-                          meterProvider: StableMeterProvider,
-                          aggregationTemporalitySelector: AggregationTemporalitySelector =
-                            AggregationTemporality.alwaysCumulative(),
-                          defaultAggregationSelector: DefaultAggregationSelector = AggregationSelector
-                            .instance,
-                          useSession: URLSession? = nil,
-                          envVarHeaders: [(String, String)]? = EnvVarHeaders.attributes) {
-    self.init(endpoint: endpoint,
-              config: config,
-              aggregationTemporalitySelector: aggregationTemporalitySelector,
-              defaultAggregationSelector: defaultAggregationSelector,
-              useSession: useSession,
-              envVarHeaders: envVarHeaders)
-    exporterMetrics = ExporterMetrics(type: "metric",
-                                      meterProvider: meterProvider,
-                                      exporterName: "otlp",
-                                      transportName: config.exportAsJson
-                                        ? ExporterMetrics.TransporterType.httpJson
-                                        : ExporterMetrics.TransporterType.grpc)
   }
 
   // MARK: - StableMetricsExporter
