+++ Sources/Exporters/OpenTelemetryProtocolHttp/trace/OtlpHttpTraceExporter.swift
@@ -20,7 +20,7 @@ public class OtlpHttpTraceExporter: OtlpHttpExporterBase, SpanExporter {
   var pendingSpans: [SpanData] = []
 
   private let exporterLock = Lock()
-  private var exporterMetrics: ExporterMetrics?
+  internal var exporterMetrics: IExporterMetrics?
 
   override
   public init(endpoint: URL = defaultOltpHttpTracesEndpoint(),
@@ -31,28 +31,6 @@ public class OtlpHttpTraceExporter: OtlpHttpExporterBase, SpanExporter {
                config: config,
                useSession: useSession,
                envVarHeaders: envVarHeaders)
-  }
-
-  /// A `convenience` constructor to provide support for exporter metric using`StableMeterProvider` type
-  /// - Parameters:
-  ///    - endpoint: Exporter endpoint injected as dependency
-  ///    - config: Exporter configuration including type of exporter
-  ///    - meterProvider: Injected `StableMeterProvider` for metric
-  ///    - useSession: Overridden `URLSession` if any
-  ///    - envVarHeaders: Extra header key-values
-  public convenience init(endpoint: URL,
-                          config: OtlpConfiguration,
-                          meterProvider: StableMeterProvider,
-                          useSession: URLSession? = nil,
-                          envVarHeaders: [(String, String)]? = EnvVarHeaders.attributes) {
-    self.init(endpoint: endpoint, config: config, useSession: useSession,
-              envVarHeaders: envVarHeaders)
-    exporterMetrics = ExporterMetrics(type: "span",
-                                      meterProvider: meterProvider,
-                                      exporterName: "otlp",
-                                      transportName: config.exportAsJson
-                                        ? ExporterMetrics.TransporterType.httpJson
-                                        : ExporterMetrics.TransporterType.grpc)
   }
 
   public func export(spans: [SpanData], explicitTimeout: TimeInterval? = nil)
