+++ Sources/OpenTelemetryApi/Propagation/ContextPropagators.swift
@@ -9,3 +9,20 @@ public protocol ContextPropagators {
   var textMapPropagator: TextMapPropagator { get }
   var textMapBaggagePropagator: TextMapBaggagePropagator { get }
 }
+
+extension OpenTelemetry {
+
+    public static func registerDefaultPropagators() {
+        self.registerPropagators(textPropagators: [W3CTraceContextPropagator()], baggagePropagator: W3CBaggagePropagator())
+    }
+
+    public static var propagators: ContextPropagators {
+        instance._propagators as! ContextPropagators
+    }
+
+    public static func registerPropagators(textPropagators: [TextMapPropagator],
+                                           baggagePropagator: TextMapBaggagePropagator) {
+      instance._propagators = DefaultContextPropagators(textPropagators: textPropagators, baggagePropagator: baggagePropagator)
+    }
+
+}
